requirements.txt

python-decouple==3.8
requests==2.31.0
requests-cache==1.1.0
requests-file==2.0.0
spotipy==2.23.0
